backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.1
pydantic==2.5.0
pydantic-settings==2.1.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
alembic==1.13.0
passlib==1.7.4
python-jose==3.3.0
python-magic==0.4.27
tensorflow==2.15.0
segment-anything==1.0
django==4.2.8
volcengine==1.0.64
pymysql==1.1.0
cryptography==41.0.7
